backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.1
pandas==2.2.2
numpy==2.3.2
scikit-learn==1.7.1
joblib==1.5.2
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
Werkzeug>=3.1.0
gunicorn